refactor(slice-11): extract remaining connectshyft orchestration handlers
