+++ apps/connectshyft-api/src/modules/connectshyft/escalationRecipientDirectory.ts
@@ -1,0 +1,331 @@
+import { Request } from 'express';
+import {
+  connectShyftEscalationRecipientScopes,
+  createEscalationRecipientDirectory,
+  type ConnectShyftEscalationRecipientDirectory,
+  type ConnectShyftEscalationRecipientOption,
+} from './escalationConfig';
+import { isConnectShyftTestOverrideEnabled } from './featureFlags';
+import { loadConnectShyftPlatformDb } from './http/accessContext';
+
+const TEST_RECIPIENT_DIRECTORY_HEADER = 'x-test-connectshyft-recipient-directory';
+const UUID_PATTERN = /^[0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$/i;
+
+const normalizeNonEmptyString = (value: unknown): string | null => {
+  if (typeof value !== 'string') {
+    return null;
+  }
+
+  const normalized = value.trim();
+  return normalized.length > 0 ? normalized : null;
+};
+
+const parseScopedRecipientEntries = (
+  rawEntries: unknown,
+  scope: ConnectShyftEscalationRecipientOption['scope'],
+): {
+  recipientIds: string[];
+  options: ConnectShyftEscalationRecipientOption[];
+} => {
+  if (!Array.isArray(rawEntries)) {
+    return {
+      recipientIds: [],
+      options: [],
+    };
+  }
+
+  const recipientIds: string[] = [];
+  const options: ConnectShyftEscalationRecipientOption[] = [];
+
+  rawEntries.forEach((entry) => {
+    if (typeof entry === 'string') {
+      const userId = normalizeNonEmptyString(entry);
+      if (!userId) {
+        return;
+      }
+
+      recipientIds.push(userId);
+      options.push({
+        value: userId,
+        label: userId,
+        scope,
+      });
+      return;
+    }
+
+    if (!entry || typeof entry !== 'object') {
+      return;
+    }
+
+    const candidate = entry as {
+      userId?: unknown;
+      label?: unknown;
+    };
+
+    const userId = normalizeNonEmptyString(candidate.userId);
+    if (!userId) {
+      return;
+    }
+
+    const label = normalizeNonEmptyString(candidate.label) || userId;
+    recipientIds.push(userId);
+    options.push({
+      value: userId,
+      label,
+      scope,
+    });
+  });
+
+  return {
+    recipientIds,
+    options,
+  };
+};
+
+const parseRecipientOptions = (
+  rawOptions: unknown,
+): ConnectShyftEscalationRecipientOption[] => {
+  if (!Array.isArray(rawOptions)) {
+    return [];
+  }
+
+  return rawOptions
+    .map((entry) => {
+      if (!entry || typeof entry !== 'object') {
+        return null;
+      }
+
+      const candidate = entry as {
+        value?: unknown;
+        label?: unknown;
+        scope?: unknown;
+      };
+
+      const value = normalizeNonEmptyString(candidate.value);
+      if (!value) {
+        return null;
+      }
+
+      const label = normalizeNonEmptyString(candidate.label) || value;
+      const scope = normalizeNonEmptyString(candidate.scope)
+        || connectShyftEscalationRecipientScopes.TEST_ONLY;
+
+      return {
+        value,
+        label,
+        scope,
+      } as ConnectShyftEscalationRecipientOption;
+    })
+    .filter((entry): entry is ConnectShyftEscalationRecipientOption => entry !== null);
+};
+
+const buildDefaultTestRecipientDirectory = (): ConnectShyftEscalationRecipientDirectory =>
+  createEscalationRecipientDirectory({
+    orgUnitRecipientIds: [
+      'user-connectshyft-a4-primary-recipient',
+      'user-connectshyft-a4-secondary-recipient',
+      'user-connectshyft-a5-orgunit-admin',
+      'user-connectshyft-a5-orgunit-member',
+    ],
+    tenantRecipientIds: [
+      'user-connectshyft-a4-primary-recipient',
+      'user-connectshyft-a4-secondary-recipient',
+      'user-connectshyft-a4-tenant-staff-recipient',
+      'user-connectshyft-a5-orgunit-admin',
+      'user-connectshyft-a5-orgunit-member',
+      'user-connectshyft-a5-tenant-staff',
+    ],
+    options: [
+      {
+        value: 'user-connectshyft-a4-primary-recipient',
+        label: 'Primary OrgUnit Admin',
+        scope: connectShyftEscalationRecipientScopes.ORG_UNIT,
+      },
+      {
+        value: 'user-connectshyft-a4-secondary-recipient',
+        label: 'Secondary OrgUnit Admin',
+        scope: connectShyftEscalationRecipientScopes.ORG_UNIT,
+      },
+      {
+        value: 'user-connectshyft-a4-tenant-staff-recipient',
+        label: 'Tenant Staff Recipient',
+        scope: connectShyftEscalationRecipientScopes.TENANT,
+      },
+      {
+        value: 'user-connectshyft-a5-orgunit-admin',
+        label: 'A5 OrgUnit Admin',
+        scope: connectShyftEscalationRecipientScopes.ORG_UNIT,
+      },
+      {
+        value: 'user-connectshyft-a5-orgunit-member',
+        label: 'A5 OrgUnit Member',
+        scope: connectShyftEscalationRecipientScopes.ORG_UNIT,
+      },
+      {
+        value: 'user-connectshyft-a5-tenant-staff',
+        label: 'A5 Tenant Staff',
+        scope: connectShyftEscalationRecipientScopes.TENANT,
+      },
+      {
+        value: 'user-connectshyft-a4-cross-tenant-recipient',
+        label: 'Cross-tenant recipient (invalid test option)',
+        scope: connectShyftEscalationRecipientScopes.TEST_ONLY,
+      },
+    ],
+  });
+
+const resolveEscalationRecipientDirectoryFromHeader = (
+  req: Request,
+): ConnectShyftEscalationRecipientDirectory | null => {
+  if (!isConnectShyftTestOverrideEnabled()) {
+    return null;
+  }
+
+  const rawHeader = req.header(TEST_RECIPIENT_DIRECTORY_HEADER);
+  if (!rawHeader) {
+    return buildDefaultTestRecipientDirectory();
+  }
+
+  try {
+    const parsed = JSON.parse(rawHeader) as {
+      orgUnitRecipients?: unknown;
+      tenantRecipients?: unknown;
+      options?: unknown;
+    };
+
+    const orgUnitRecipients = parseScopedRecipientEntries(
+      parsed.orgUnitRecipients,
+      connectShyftEscalationRecipientScopes.ORG_UNIT,
+    );
+    const tenantRecipients = parseScopedRecipientEntries(
+      parsed.tenantRecipients,
+      connectShyftEscalationRecipientScopes.TENANT,
+    );
+
+    const options = [
+      ...orgUnitRecipients.options,
+      ...tenantRecipients.options,
+      ...parseRecipientOptions(parsed.options),
+    ];
+
+    return createEscalationRecipientDirectory({
+      orgUnitRecipientIds: orgUnitRecipients.recipientIds,
+      tenantRecipientIds: [
+        ...tenantRecipients.recipientIds,
+        ...orgUnitRecipients.recipientIds,
+      ],
+      options,
+    });
+  } catch (_error) {
+    return buildDefaultTestRecipientDirectory();
+  }
+};
+
+const buildRecipientLabel = (
+  userId: string,
+  firstName: unknown,
+  lastName: unknown,
+): string => {
+  const first = typeof firstName === 'string' ? firstName.trim() : '';
+  const last = typeof lastName === 'string' ? lastName.trim() : '';
+  const fullName = `${first} ${last}`.trim();
+  return fullName.length > 0 ? fullName : userId;
+};
+
+const buildDatabaseRecipientDirectory = async (
+  tenantId: string,
+  orgUnitId: string,
+): Promise<ConnectShyftEscalationRecipientDirectory> => {
+  if (!UUID_PATTERN.test(tenantId) || !UUID_PATTERN.test(orgUnitId)) {
+    return createEscalationRecipientDirectory({
+      orgUnitRecipientIds: [],
+      tenantRecipientIds: [],
+      options: [],
+    });
+  }
+
+  const db = loadConnectShyftPlatformDb();
+
+  const tenantRows = await db('platform.tenant_memberships as tm')
+    .leftJoin('users as u', 'u.id', 'tm.user_id')
+    .where('tm.tenant_id', tenantId)
+    .select('tm.user_id as userId', 'u.first_name as firstName', 'u.last_name as lastName');
+
+  const orgUnitRows = await db('platform.org_unit_memberships as om')
+    .join('platform.org_units as ou', 'ou.id', 'om.org_unit_id')
+    .leftJoin('users as u', 'u.id', 'om.user_id')
+    .where('om.org_unit_id', orgUnitId)
+    .andWhere('ou.tenant_id', tenantId)
+    .select('om.user_id as userId', 'u.first_name as firstName', 'u.last_name as lastName');
+
+  const orgUnitRecipientIds: string[] = [];
+  const orgUnitOptions: ConnectShyftEscalationRecipientOption[] = [];
+  orgUnitRows.forEach((row) => {
+    const userId = normalizeNonEmptyString((row as { userId?: unknown }).userId);
+    if (!userId) {
+      return;
+    }
+
+    orgUnitRecipientIds.push(userId);
+    orgUnitOptions.push({
+      value: userId,
+      label: buildRecipientLabel(
+        userId,
+        (row as { firstName?: unknown }).firstName,
+        (row as { lastName?: unknown }).lastName,
+      ),
+      scope: connectShyftEscalationRecipientScopes.ORG_UNIT,
+    });
+  });
+
+  const tenantRecipientIds: string[] = [];
+  const tenantOptions: ConnectShyftEscalationRecipientOption[] = [];
+  tenantRows.forEach((row) => {
+    const userId = normalizeNonEmptyString((row as { userId?: unknown }).userId);
+    if (!userId) {
+      return;
+    }
+
+    tenantRecipientIds.push(userId);
+    tenantOptions.push({
+      value: userId,
+      label: buildRecipientLabel(
+        userId,
+        (row as { firstName?: unknown }).firstName,
+        (row as { lastName?: unknown }).lastName,
+      ),
+      scope: connectShyftEscalationRecipientScopes.TENANT,
+    });
+  });
+
+  const directory = createEscalationRecipientDirectory({
+    orgUnitRecipientIds,
+    tenantRecipientIds,
+    options: [...orgUnitOptions, ...tenantOptions],
+  });
+
+  directory.options.sort((a, b) => {
+    if (a.label < b.label) {
+      return -1;
+    }
+    if (a.label > b.label) {
+      return 1;
+    }
+    return a.value.localeCompare(b.value);
+  });
+
+  return directory;
+};
+
+export const resolveConnectShyftEscalationRecipientDirectory = async (
+  req: Request,
+  tenantId: string,
+  orgUnitId: string,
+): Promise<ConnectShyftEscalationRecipientDirectory> => {
+  const testDirectory = resolveEscalationRecipientDirectoryFromHeader(req);
+  if (testDirectory) {
+    return testDirectory;
+  }
+
+  return buildDatabaseRecipientDirectory(tenantId, orgUnitId);
+};
