+++ apps/connectshyft-api/src/modules/connectshyft/handlers/index.ts
@@ -2,19 +2,27 @@ export { getConnectAvailability } from './getConnectAvailability';
 export { getConnectNeighborDetail } from './getConnectNeighborDetail';
 export { getConnectNeighbors } from './getConnectNeighbors';
 export { getConnectContext } from './getConnectContext';
+export { getConnectEscalationConfig } from './getConnectEscalationConfig';
+export { getConnectEscalationRecipients } from './getConnectEscalationRecipients';
 export { getConnectInbox } from './getConnectInbox';
+export { getConnectNumberMappings } from './getConnectNumberMappings';
 export { getConnectSettingsNavigation } from './getConnectSettingsNavigation';
 export { getConnectThreadDetail } from './getConnectThreadDetail';
 export { getConnectThreadTimeline } from './getConnectThreadTimeline';
+export { getConnectWebhookReceiptMetrics } from './getConnectWebhookReceiptMetrics';
 export { deleteConnectNeighbor } from './deleteConnectNeighbor';
 export { postConnectNeighborCreate } from './postConnectNeighborCreate';
 export { postConnectNeighborIdentityMatch } from './postConnectNeighborIdentityMatch';
 export { postConnectNeighborMerge } from './postConnectNeighborMerge';
+export { postConnectNumberMapping } from './postConnectNumberMapping';
 export { postConnectThreadClaim } from './postConnectThreadClaim';
 export { postConnectThreadCall } from './postConnectThreadCall';
 export { postConnectThreadClose } from './postConnectThreadClose';
 export { postConnectThreadMessage } from './postConnectThreadMessage';
 export { postConnectThreadTakeover } from './postConnectThreadTakeover';
+export { postConnectWebhookReceiptCleanup } from './postConnectWebhookReceiptCleanup';
 export { postConnectWebhookInbound } from './postConnectWebhookInbound';
 export { postConnectWebhookSms } from './postConnectWebhookSms';
 export { putConnectNeighbor } from './putConnectNeighbor';
+export { putConnectEscalationConfig } from './putConnectEscalationConfig';
+export { putConnectNumberMapping } from './putConnectNumberMapping';
